Added two new resources
I came across this repository and saw that you had linked to my JavaScript Operators chapter, which I really appreciate, and so thought of adding two further links in the JavaScript Crash Course and JavaScript Exercises sections. The JavaScript course at CodeGuage contains 44 exercises as of now, and I am working on adding more and more.